added personal preferences page
