+++ larpmanager/forms/association.py
@@ -37,6 +37,7 @@ from larpmanager.forms.utils import (
 )
 from larpmanager.models.access import AssocPermission, AssocRole
 from larpmanager.models.association import Association, AssocText, AssocTextType
+from larpmanager.models.member import Member
 
 
 class ExeAssociationForm(MyForm):
@@ -274,10 +275,6 @@ class ExeConfigForm(ConfigForm):
         label = _("Old interface")
         help_text = _("If checked: uses old interface")
         self.add_configs("interface_old", ConfigType.BOOL, label, help_text)
-
-        label = _("Expand sidebar")
-        help_text = _("If checked: automatically expand all sidebars")
-        self.add_configs("interface_expand_sidebar", ConfigType.BOOL, label, help_text)
 
         # CALENDAR
         self.set_section("calendar", _("Calendar"))
@@ -643,3 +640,24 @@ class ExeQuickSetupForm(QuickSetupForm):
             )
 
         self.init_fields(get_assoc_features(self.instance.pk))
+
+
+class ExePreferencesForm(ConfigForm):
+    page_title = _("Personal preferences")
+
+    page_info = _("This page allows you to set your personal preferences on the interface")
+
+    class Meta:
+        model = Member
+        fields = ()
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.prevent_canc = True
+        self.show_sections = True
+
+    def set_configs(self):
+        self.set_section("interface", _("Interface"))
+        label = _("Collapse sidebar")
+        help_text = _("If checked: collpase sidebars links, and expand on mouse hover")
+        self.add_configs("interface_collapse_sidebar", ConfigType.BOOL, label, help_text)
