+++ larpmanager/locale/fr/LC_MESSAGES/django.po
@@ -2105,9 +2105,6 @@ msgstr "Caractéristiques de l'événement"
 msgid "Executive Vote"
 msgstr "Vote de la direction"
 
-msgid "Expand sidebar"
-msgstr "Agrandir la barre latérale"
-
 msgid "Form questions"
 msgstr "Questions de forme"
 
@@ -2371,6 +2368,9 @@ msgstr "Priorité au casting"
 
 msgid "Character review"
 msgstr "Revue des personnages"
+
+msgid "Collapse sidebar"
+msgstr "Réduire la barre latérale"
 
 msgid "Collection links"
 msgstr "Liens de la collection"
@@ -3857,6 +3857,9 @@ msgstr "Sélectionnez la nouvelle photo à télécharger"
 msgid "Set up the taxes configuration"
 msgstr "Configuration des taxes"
 
+msgid "Set your interface preferences"
+msgstr "Définissez vos préférences en matière d'interface"
+
 msgid "The registrations to the event"
 msgstr "Inscription à l'événement"
 
@@ -4908,9 +4911,6 @@ msgstr "(Facultatif) Le montant d'un don volontaire"
 msgid "Customize buttons in the event navigation bar"
 msgstr "Personnaliser les boutons de la barre de navigation de l'événement"
 
-msgid "If checked: automatically expand all sidebars"
-msgstr "Si coché : développer automatiquement toutes les barres latérales"
-
 msgid "If checked: shows the position for each event"
 msgstr "Si coché : indique la position pour chaque événement"
 
@@ -5000,9 +5000,6 @@ msgstr "Les utilisateurs enregistrés bénéficient d'un accès normal"
 
 msgid "Remember to put recipients separated by commas"
 msgstr "N'oubliez pas de séparer les destinataires par des virgules"
-
-msgid "Set your personal preferences on the interface"
-msgstr "Définissez vos préférences personnelles sur l'interface"
 
 msgid "Something went wrong in the account activation"
 msgstr "Un problème s'est produit lors de l'activation du compte"
@@ -6003,6 +6000,11 @@ msgid "From this page you can send an email to any list of recipients"
 msgstr ""
 "À partir de cette page, vous pouvez envoyer un courrier électronique à "
 "n'importe quelle liste de destinataires"
+
+msgid "If checked: collpase sidebars links, and expand on mouse hover"
+msgstr ""
+"Si la case est cochée : les liens de la barre latérale sont masqués et "
+"développés au survol de la souris"
 
 msgid "If you encounter any issues, contact us and we will assist you"
 msgstr ""
