+++ larpmanager/locale/pl/LC_MESSAGES/django.po
@@ -2105,9 +2105,6 @@ msgstr "Cechy wydarzenia"
 msgid "Executive Vote"
 msgstr "Głosowanie wykonawcze"
 
-msgid "Expand sidebar"
-msgstr "Rozwiń pasek boczny"
-
 msgid "Form questions"
 msgstr "Pytania dotyczące formularza"
 
@@ -2371,6 +2368,9 @@ msgstr "Priorytet odlewania"
 
 msgid "Character review"
 msgstr "Przegląd postaci"
+
+msgid "Collapse sidebar"
+msgstr "Zwiń pasek boczny"
 
 msgid "Collection links"
 msgstr "Linki do kolekcji"
@@ -3857,6 +3857,9 @@ msgstr "Wybierz nowe zdjęcie do przesłania"
 msgid "Set up the taxes configuration"
 msgstr "Konfiguracja podatków"
 
+msgid "Set your interface preferences"
+msgstr "Ustaw preferencje interfejsu"
+
 msgid "The registrations to the event"
 msgstr "Rejestracje na wydarzenie"
 
@@ -4900,9 +4903,6 @@ msgstr "(Opcjonalnie) Kwota dobrowolnej darowizny"
 msgid "Customize buttons in the event navigation bar"
 msgstr "Dostosowywanie przycisków na pasku nawigacji wydarzeń"
 
-msgid "If checked: automatically expand all sidebars"
-msgstr "Jeśli zaznaczone: automatycznie rozwija wszystkie paski boczne"
-
 msgid "If checked: shows the position for each event"
 msgstr "Jeśli zaznaczone: pokazuje pozycję dla każdego zdarzenia"
 
@@ -4990,9 +4990,6 @@ msgstr "Zarejestrowani użytkownicy otrzymają normalny dostęp"
 
 msgid "Remember to put recipients separated by commas"
 msgstr "Pamiętaj, aby oddzielić odbiorców przecinkami"
-
-msgid "Set your personal preferences on the interface"
-msgstr "Ustaw osobiste preferencje w interfejsie"
 
 msgid "Something went wrong in the account activation"
 msgstr "Coś poszło nie tak podczas aktywacji konta"
@@ -5922,6 +5919,10 @@ msgstr "Opis tej metody płatności wyświetlany użytkownikowi"
 msgid "From this page you can send an email to any list of recipients"
 msgstr ""
 "Z tej strony można wysłać wiadomość e-mail do dowolnej listy odbiorców"
+
+msgid "If checked: collpase sidebars links, and expand on mouse hover"
+msgstr ""
+"Jeśli zaznaczone: zwija linki pasków bocznych i rozwija po najechaniu myszą"
 
 msgid "If you encounter any issues, contact us and we will assist you"
 msgstr ""
