+++ larpmanager/templates/elements/structure/sidebar.html
@@ -26,6 +26,14 @@
                         <a href="#" class="explain-icon" tog="manage"><i class="fas fa-question-circle"></i></a>
                     </span>
                 </p>
+                <p>
+                    <span class="sidebar-link-wrap"
+                          descr="{% trans "Set your interface preferences" %}">
+                        <a class="sidebar-link"
+                           href="{% url 'orga_preferences' run.event.slug run.number %}">{% trans "Preferences" %}</a>
+                        <a href="#" class="explain-icon" tog="orga_preferences"><i class="fas fa-question-circle"></i></a>
+                    </span>
+                </p>
                 {% for mod, list in event_pms.items %}
                     {% if list %}
                         <div class="sidebar-section">
@@ -55,6 +63,14 @@
                           descr="{% trans "Access the organization dashboard" %}">
                         <a class="sidebar-link" href="{% url 'manage' %}">Dashboard</a>
                         <a href="#" class="explain-icon" tog="manage"><i class="fas fa-question-circle"></i></a>
+                    </span>
+                </p>
+                <p>
+                    <span class="sidebar-link-wrap"
+                          descr="{% trans "Set your interface preferences" %}">
+                        <a class="sidebar-link"
+                           href="{% url 'exe_preferences' run.event.slug run.number %}">{% trans "Preferences" %}</a>
+                        <a href="#" class="explain-icon" tog="exe_preferences"><i class="fas fa-question-circle"></i></a>
                     </span>
                 </p>
                 {% for mod, list in assoc_pms.items %}
