+++ larpmanager/templates/trans.html
@@ -436,7 +436,5 @@
 {% blocktrans %}Display sent e-mails{% endblocktrans %}
 {% blocktrans %}Quick Setup{% endblocktrans %}
 {% blocktrans %}quick{% endblocktrans %}
-{% blocktrans %}Preferences{% endblocktrans %}
-{% blocktrans %}Set your personal preferences on the interface{% endblocktrans %}
 {% blocktrans %}Reading{% endblocktrans %}
 {% blocktrans %}Read the writing elements{% endblocktrans %}
